refactor(#164): annotations/turned.py + _common.py (P5b)
Continue the annotations/ split. P5b moves the turned-part diameter callouts
(_annotate_turned_diameters + its turned-only helpers _mentioned_diams,
_distinct_bosses, _turned_diameters_beside) into annotations/turned.py.

The placement helpers _anno_box / _occupied_boxes / _box_hits are shared by the
turned pass AND the holes pass, so they move to annotations/_common.py (the bottom
of the annotations DAG) rather than to either pass. turned.py imports them from
_common; annotate.py re-imports both _common helpers (still used by the holes
orchestrator) and _annotate_turned_diameters (called by _auto_annotate) — all
downward, no cycle.

The turned monkeypatch test now patches the strip solvers on annotations.turned
(where the pass resolves them). Pure move: golden gate passes against the committed
snapshots with no regeneration; full make_drawing/golden suites pass (357); ruff
check + format + mypy clean. Adversarially checked: no upward imports, no stale
defs, turned parts get diameter callouts and holed parts still place clear.

Part of #164 (P5c-P5e follow).
